requirements.txt
PyYAML
Flask==1.1.1
nboost[pt]
elasticsearch